mutil-api/requirements.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0
requests==2.28.1
openai==1.3.0
python-dotenv==0.21.0
python-telegram-bot==20.6